Removed JSONUtil
JSONUtil had no real responsibilities, the relevant methods of JSONUtil
have been moved to the EntityUtil.
